refactor(operator_commands_ooda): split daemon.rs into mod+helpers+config (#1266)
daemon.rs 478 -> daemon/{mod.rs 391, helpers.rs 74, config.rs 21}.
